remove background process for upload
