+++ package.json
@@ -11,7 +11,6 @@
     "url": "http://github.com/kevinhughes27/parity-server"
   },
   "dependencies": {
-    "background-process": "^0.1.0",
     "body-parser": "^1.13.3",
     "dotenv": "^1.2.0",
     "express": "^4.14.0",
